Handle potential API errors

+++ src/main/java/br/com/dv/integracaoapitransport/app/model/fastcommerce/request/FreightQuoteRequest.java
@@ -1,11 +1,12 @@
 package br.com.dv.integracaoapitransport.app.model.fastcommerce.request;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record FreightQuoteRequest(
         String fromCEP,
         String toCEP,
-        double totalAmount,
+        BigDecimal totalAmount,
         double totalWeight,
         boolean hasBulky,
         CalculationParameters calculationParameters,
