+++ src/main/java/br/com/dv/integracaoapitransport/app/model/fastcommerce/response/ShippingService.java
@@ -1,4 +1,6 @@
 package br.com.dv.integracaoapitransport.app.model.fastcommerce.response;
 
-public record ShippingService(String serviceName, double servicePrice, String serviceNotes) {
+import java.math.BigDecimal;
+
+public record ShippingService(String serviceName, BigDecimal servicePrice, String serviceNotes) {
 }
