+++ src/main/java/br/com/dv/integracaoapitransport/app/model/transport/carriers/response/Frete.java
@@ -1,4 +1,6 @@
 package br.com.dv.integracaoapitransport.app.model.transport.carriers.response;
 
-public record Frete(String modalidade, int prazo, double preco) {
+import java.math.BigDecimal;
+
+public record Frete(String modalidade, Integer prazo, BigDecimal preco) {
 }
