+++ src/main/java/br/com/dv/integracaoapitransport/app/model/transport/jadlog/response/FreteResponse.java
@@ -1,19 +1,21 @@
 package br.com.dv.integracaoapitransport.app.model.transport.jadlog.response;
 
+import java.math.BigDecimal;
+
 public record FreteResponse(
         String cepori,
         String cepdes,
         String frap,
-        double peso,
+        Double peso,
         String cnpj,
         String conta,
         String contrato,
-        int modalidade,
+        Integer modalidade,
         String tpentrega,
         String tpseguro,
-        double vldeclarado,
-        double vlcoleta,
-        int prazo,
-        double vltotal
+        BigDecimal vldeclarado,
+        BigDecimal vlcoleta,
+        Integer prazo,
+        BigDecimal vltotal
 ) {
 }
